requirements.txt
openai
streamlit
python-dotenv
pandas
